+++ src/commands/runMacro.ts
@@ -1,13 +1,11 @@
 import * as vscode from 'vscode';
-import { MACROS_FILTER } from '../common/ui';
+import { showMacroOpenDialog } from '../common/ui';
 import { Manager } from '../manager';
 
 export async function runMacro(manager: Manager, pathOrUri?: string | vscode.Uri) {
   const targetUri = pathOrUri
     ? (pathOrUri instanceof vscode.Uri ? pathOrUri : vscode.Uri.file(pathOrUri))
-    : (await vscode.window.showOpenDialog({
-      filters: MACROS_FILTER,
-    }))?.pop();
+    : await showMacroOpenDialog();
 
   if (!targetUri) {
     return; // Nothing to run.