+++ src/extension.ts
@@ -1,9 +1,10 @@
 import * as vscode from 'vscode';
 import { createMacro } from './commands/createMacro';
-import { runMacroFromSourceDirs } from './commands/runFromSourceDirs';
+import { debugMacro } from './commands/debugMacro';
+import { getActiveEditorUri } from './commands/getActiveEditorUri';
+import { getMacroFromSourceDirs } from './commands/getMacroFromSourceDirs';
 import { runMacro } from './commands/runMacro';
-import { showRunningMacros } from './commands/showRunninMacros';
-import { MACROS_FILTER } from './common/ui';
+import { showRunningMacros } from './commands/showRunningMacros';
 import { Manager } from './manager';
 import { StatusBarItem } from './statusBarItem';
 
@@ -23,6 +24,19 @@ export async function activate(context: vscode.ExtensionContext) {
   context.subscriptions.push(
     manager,
     new StatusBarItem(manager),
+    vscode.commands.registerCommand('macros.debug', (pathOrUri?: string | vscode.Uri) => debugMacro(manager, pathOrUri)),
+    vscode.commands.registerCommand('macros.debug.activeEditor', async () => {
+      const uri = await getActiveEditorUri();
+      if (uri) {
+        await debugMacro(manager, uri);
+      }
+    }),
+    vscode.commands.registerCommand('macros.debug.fromSourceDirs', async () => {
+      const uri = await getMacroFromSourceDirs();
+      if (uri) {
+        await debugMacro(manager, uri);
+      }
+    }),
     vscode.commands.registerCommand('macros.new.macro', () => createMacro(context)),
     vscode.commands.registerCommand('macros.run', (pathOrUri?: string | vscode.Uri) => runMacro(manager, pathOrUri)),
     vscode.commands.registerCommand('macros.run.activeEditor', async () => {
@@ -31,37 +45,14 @@ export async function activate(context: vscode.ExtensionContext) {
         await runMacro(manager, uri);
       }
     }),
-    vscode.commands.registerCommand('macros.run.fromSourceDirs', () => runMacroFromSourceDirs()),
+    vscode.commands.registerCommand('macros.run.fromSourceDirs', async () => {
+      const uri = await getMacroFromSourceDirs();
+      if (uri) {
+        await runMacro(manager, uri);
+      }
+    }),
     vscode.commands.registerCommand('macros.run.mru', () => runMacro(manager, mruMacro)),
     vscode.commands.registerCommand('macros.run.show', () => showRunningMacros(manager)),
   );
-}
-
-async function getActiveEditorUri(): Promise<vscode.Uri | undefined> {
-  let uri: vscode.Uri | undefined;
-  const editor = vscode.window.activeTextEditor;
-  if (editor) {
-    const { document } = editor;
-    if (document.isUntitled) {
-      uri = await saveUntitled(document);
-    } else if (!document.isDirty || await document.save()) {
-      uri = document.uri;
-    }
-  }
-
-  return uri;
-
-  async function saveUntitled(document: vscode.TextDocument) {
-    const targetUri = await vscode.window.showSaveDialog({
-      filters: MACROS_FILTER
-    });
-
-    if (targetUri) {
-      await vscode.workspace.fs.writeFile(targetUri, Buffer.from(document.getText()));
-      await vscode.commands.executeCommand('workbench.action.revertAndCloseActiveEditor');
-      await vscode.window.showTextDocument(targetUri, { preview: false });
-    }
-    return targetUri;
-  }
 }